Put specs into new file & fix first error
Move the tests into their own spec file so it is easier to work with them.
Additionally, add a fail statement to the initialize method for the
ConferenceTicketOrder class so that it will fail if a user requests anything
other than one ticket in the order. This allows the first of two failing
tests to pass.