Refactor shipping_cost and shipping variable
The shipping_cost method was setting a local variable 'shipping' equal
to zero which wasn't being used anywhere. Since it only needs to return
the number zero, the local variable assignment was removed.

Additionally, since this method can be referenced from anywhere within
the class, the assignment to the shipping var in the to_s method was
removed and the new method is now referenced from the private total method.