Extract Class: Payment
The charge method was doing a lot, include processing a payment,
sending an email, and updating the status to equal 'charged.' If
new forms of payment are going to be added in the future, it makes
sense to extract some of this logic into its own class.

This commit creates the Payment class which is responsible for processing
any payments, regardless of payment type. It is initialized with the payment
type, then its main method "process" handles the actual processing of the
payments.