Extract Class: Order Report
Because there may need to be different types of report output formats
in the future, I extracted out a class for OrderReport. When formats
are added in the future, additional methods can simply be added to this
class, as can additional branches of the case statement to accommodate the
changes. This further reduces the responsibility of the ConferenceTicketOrder
class. Also, the OrderReport class could be used with any other type of order.